CY-1891 Make watch handlers only acquire locks when necessary (#131)
Watch handlers used to acquire the lock over the whole run method,
which also included fetching data. However, fetching is a long-polling
operation, which only returns when the data actually changed, and
so the lock was unnecessarily held almost all the time.

The lock is only necessary around actual running of the handlers, not
around long-polling for data.

See the description of ticket CY-1891 for notes about impact of this
issue.